[fix] #57 meter , page 타입 변경 Long -> Int

+++ data/src/main/java/com/dpm/sixpack/data/source/remote/dto/response/FeedResponseDto.kt
@@ -103,9 +103,9 @@ data class FeedDto(
             selfieTime = selfieTime,
             runningSessionResult =
                 RunningSessionResult(
-                    totalDistanceMeter = totalDistance,
+                    totalDistanceMeter = totalDistance.toInt(),
                     totalDurationSec = totalRunTime,
-                    avgPace = averagePace,
+                    avgPace = averagePace.toInt(),
                     avgCadence = cadence,
                 ),
             imageUrl = imageUrl,
